client: change no_delay flag to a flags field
In a later patch, we'll want to have the client set the sync flag in
the cap flush, to hint to the MDS that it should process it immediately.

We could add a second bool, but let's instead do what the kernel client
does which is to have a flags field. With that, the existing no_delay
bool becomes CHECK_CAPS_NODELAY.

We'll add other flags in subsequent patches.

Signed-off-by: Jeff Layton <jlayton@redhat.com>